// src/enums/permission.enum.ts
export enum Permission {
  VIEW_EMPLOYEE = "view_employee",
  CREATE_EMPLOYEE = "create_employee",
  DELETE_EMPLOYEE = "delete_employee",
  UPDATE_EMPLOYEE = "update_employee",

  CREATE_CUSTOMER = "create_customer",
  VIEW_CUSTOMER = "view_customer",
  DELETE_CUSTOMER = "delete_customer",
  UPDATE_CUSTOMER = "update_customer",
  CUSTOMER_CREATE_MANAGER = "customer_create_manager",

  CREATE_CONTRACT = "create_contract",
  VIEW_CONTRACT = "view_contract",
  DELETE_CONTRACT = "delete_contract",
  UPDATE_CONTRACT = "update_contract",
  CONTRACT_CREATE_MANAGER = "contract_create_manager",

  CREATE_DEBTOR = "create_debtor",
  VIEW_DEBTOR = "view_debtor",
  DELETE_DEBTOR = "delete_debtor",
  UPDATE_DEBTOR = "update_debtor",

  CREATE_CASH = "create_cash",
  VIEW_CASH = "view_cash",
  DELETE_CASH = "delete_cash",
  UPDATE_CASH = "update_cash",

  CREATE_PAYMENT = "create_payment",
  VIEW_PAYMENT = "view_payment",
  DELETE_PAYMENT = "delete_payment",
  UPDATE_PAYMENT = "update_payment",

  CREATE_EXPENSE = "create_expense",
  VIEW_EXPENSE = "view_expense",
  DELETE_EXPENSE = "delete_expense",
  UPDATE_EXPENSE = "update_expense",

  VIEW_DASHBOARD = "view_dashboard",
  VIEW_REPORTS = "view_reports",
}